+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Admin Cost Center addresses UI
+
 ## [1.36.0] - 2024-10-16
 
 ### Added
@@ -14,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Full Impersonation` Organization Setting, this will allow a Impersonator to also Switch between the User's Organization/Cost Center
 
 ## [1.35.2] - 2024-10-15
+
 ### Added
 
 - Add new tab called 'settings' at organization details